Allow creating new Resources
This also surfaced a bug where SWR would think data hadn't changed,
because it's doing a deep comparison of the SolidDataset objects,
not the contents of the Dataset. As a workaround for now, I've
implemented my own comparison function that simply serialises them
and compares the serialisation, but of course that's even worse for
performance than a deep comparison. Hopefully I can get permission
to switch solid-client to plain objects ASAP :)