Upgrade SCAB, disable auto-session restore
The experience of restoring the session automatically unfortunately
is too jarring (a redirection away and back just after the page has
loaded), so I've explicitly disabled it in case it gets enabled by
default again in the future.
I did already add the code to restore the route after restoring the
session, just in case we need to be ready for that.