Only mark self-contained items as self-contained
The "Thing" element wasn't actually self-contained, since the
deletion button escaped its boundaries. Thus, we now only mark the
content of the Thing as self-contained, whereas the larger wrapper
(including the heading) can be broken out of by the deletion
button.

This also fixes a bug where the deletion button was hidden when the
content was collapsed, because it was part of the content's DOM
structure.