Allow forcing an Accept header for useResource
Generally we don't know what Content Type the user will be viewing,
but in some cases, we know we'll want to request `text/turtle` for
a Resource that might also default to a different Content-Type if
no Accept header is passed. (Which is kind of annoying, because
browsers do pass an Accept header as far as I know, so they could
just default to the user-friendly representation when that is
present.)

Specifically, Inrupt's Enterprise Solid Server has started serving
HTML for WebIDs.