Don't forget about the root while data is loading
When switching Resources with the same root, if the Resource is
still loading but the URL is already known, we might still have
URL of the root cached. Thus, when switching Resources, we
should fall back to the Resource URL to determine the root while
awaiting the Resource to load, otherwise the tree view might
temporarily disappear while we wait for the Resource to load.